# fix(bandersnatch): GLV bounds + test