… internal/generator/edwards/template/point.go.tmpl
… internal/generator/edwards/template/tests/point.go.tmpl