# Description

When we revisited GLV scalar decomposition https://github.com/Consensys/gnark-crypto/pull/213, we increased conditionally the loop bounds by 1 for all `mulGLV` methods for all curves but we forgot to do so for Bandersnatch twisted Edwards curve. This PR fixes the issue. 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How has this been tested?

added `scalarMulWindowed` which does a windowed double-and-add and compares the result with `scalarMulGLV` which uses the endomorphism. The method `ScalarMultiplication` picks `scalarMulGLV` whenever an efficient endomorphism is available (currently only for Bandersnatch).

# How has this been benchmarked?

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkScalarMulExtended-8       30378         30271         -0.35%
BenchmarkScalarMulProjective-8     33844         33941         +0.29%
```

No major diff for random scalars.

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules