@yelhousni 2024-07-02
@ThomasPiellard this is just consolidating GLV in twisted Edwards with Weiestrass.